fixed an issue where pipeline cannot return int64 value, the fix is apply for float also

+++ pipeline/stage_test.go
@@ -8,12 +8,12 @@ func TestNextStage(t *testing.T) {
 
 	fn := func() (struct {
 		Field string
-		Meo   int
+		Meo   int64
 	}, error) {
 
 		return struct {
 			Field string
-			Meo   int
+			Meo   int64
 		}{"meomeocute", 1}, nil
 	}
 	stage := NewStage(fn)
